Add examples, renaming and tidying
