--- etiqet-db/src/test/java/com/neueda/etiqet/db/fixture/DbServerNotFoundException.java
@@ -1,11 +1,0 @@
-package com.neueda.etiqet.db.fixture;
-
-public class DbServerNotFoundException extends RuntimeException {
-    public DbServerNotFoundException() {
-        super();
-    }
-
-    public DbServerNotFoundException(String message) {
-        super(message);
-    }
-}
