--- etiqet-db/src/test/java/com/neueda/etiqet/db/fixture/stepdefs/Initialization.java
@@ -1,12 +1,0 @@
-package com.neueda.etiqet.db.fixture.stepdefs;
-
-import com.neueda.etiqet.db.fixture.DbHandlers;
-import cucumber.api.java.en.Given;
-
-public class Initialization {
-
-    @Given("^I? ?connect to (?:the)? database$")
-    public void connect() {
-        DbHandlers.connect(DbHandlers.DEFAULT_SERVER_ALIAS);
-    }
-}
