+++ etiqet-sql/src/test/java/com/neueda/etiqet/sql/fixture/SqlServerNotFoundException.java
@@ -1,0 +1,11 @@
+package com.neueda.etiqet.sql.fixture;
+
+public class SqlServerNotFoundException extends RuntimeException {
+    public SqlServerNotFoundException() {
+        super();
+    }
+
+    public SqlServerNotFoundException(String message) {
+        super(message);
+    }
+}
