+++ etiqet-sql/src/test/java/com/neueda/etiqet/sql/fixture/stepdefs/Util.java
@@ -1,0 +1,12 @@
+package com.neueda.etiqet.sql.fixture.stepdefs;
+
+import com.neueda.etiqet.sql.fixture.SqlHandlers;
+import cucumber.api.java.en.Then;
+
+public class Util {
+
+    @Then("^I? ?print (?:the)? ?results$")
+    public void printResults() {
+        SqlHandlers.printResults();
+    }
+}
