+++ etiqet-sql/src/test/resources/scenarios/connecting.feature
@@ -1,0 +1,4 @@
+Feature: Connecting to a database
+
+    Scenario: Connect with default config
+        Given I connect to the database
